feat: moved sidebar
