+++ src/client/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import SettingsPane from './pages/settings/SettingsPane';
 import ViewPane from './pages/view/ViewPane';
 import Stack from '@mui/joy/Stack';
 import Divider from '@mui/joy/Divider';
 import AlertPopup from './modals/AlertPopup';
+import SettingsPane from './pages/settings/SettingsPane';
 
 /** ------------------------------------------------------------------------- */
