+++ src/client/sidebar/TransformerSettings.tsx
@@ -5,8 +5,8 @@ import ListItemContent from '@mui/joy/ListItemContent';
 import Typography from '@mui/joy/Typography';
 import AccordionDetails from '@mui/joy/AccordionDetails';
 import React from 'react';
-import { useAppDispatch, useAppSelector } from '../../../client/store/hooks';
-import { getTransformers, getTransformersSettings, setTransformersNames, setTransformersTags } from '../../../client/store/slices/system';
+import { useAppDispatch, useAppSelector } from '../../client/store/hooks';
+import { getTransformers, getTransformersSettings, setTransformersNames, setTransformersTags } from '../../client/store/slices/system';
 import Chip from '@mui/joy/Chip';
 import FormControl from '@mui/joy/FormControl';
 import Option from '@mui/joy/Option';
